erreur point

+++ ProjetJava2025/src/Projet/Point.java
@@ -5,11 +5,11 @@ public class Point {
 	private double y;
 	private double z;
 
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
-	public Point(int x, int y, int z) {
+	public Point(double x, double y, double z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
